ENGR00325392 [LMX6Q_SD]GST1.0 XvidDec: Video is not fluent after 13 seconds of playing. 100%
The clip is LPCM audio, no audio decoder in the pipeline, so the
pipeline buffer is low.
The VPU buffered too many buffer when latency mode. The default VPU
buffer is 3M bytes. will buffer about 700 video frame, 20+ seconds.
The resolution is control buffer count and buffer time when latency.

Signed-off-by: Song Bing b06498@freescale.com