[MMFMWK-6480] Fix GST1.4 2 mp4 stress file cannot play
The root cause of this issue is that
For these clips the parser will return seek error in the middle of the stream.
and there will be multiqueue linked after aiurdemux.
Previously in GST1.2 the multiqueue property max-size-buffers is set as default value 5, so that when multiqueue buffers over 5, it will push data downstream, and the pipeline is running.
However in gst1.4 decodebin set multi-queue max-size-buffers to 0, so that multiqueue will hold no limitation buffers until exceed max-siize-bytes(2M), but the error is within 2M data, so downstream decoder cannot get the data flow, and cannot play,

Resolution: in aiurdemux when get parser seek error, should not just return GST_FLOW_ERROR, need return GST_FLOW_EOS, and push eos event downstream, so that multi-queue can drain data to downstream, and pipeline can run.

Signed-off-by: Lyon Wang <lyon.wang@freescale.com>