[MMFMWK-6870] Glimagesink: No audio output if play with GLimagesink. 100%
Root cause:
when video sink is specified, the video sink name is stored in an gchar
array. This issue was cause by the length of this array is not enough.

Solution:
Change to use gchar pointer to store the string address of video sink name
Audio sink, text sink and suburi are also should modified.

Signed-off-by: Haihua Hu <b55597@freescale.com>