plugins: aiurdemux: aiurcontent: Add "mtp-track" as a protocol.
There are some use cases with MP4 container files over MTP, where the stream
cache causes a randomly but frequently push_seek() in order to obtain the stream
data in its local cache.  This causes the audio to be distorted due to its
frequent seek followed by gst_pad_push_event(); with GST_SEEK_FLAG_FLUSH set,
which in turn flash'es all its buffers.

Mostly, this is due to that the actual address to be seek'ed which is called in
gst_aiur_stream_cache_seek(); is out of bounds within its cache memory range which
causes the stream cache to seek by calling gst_pad_push_event();

The solution is to add "mtp-track" as a protocol support to enable FILE_FLAG_READ_IN_SEQUENCE
flags to inform the FslParser to disable random_access on its stream cache and
and read the file in sequence.

This solution is being tested and fixed the issue, with the following plugin
as an <mtp_src>:
  - WrsMtpSrc

  Also, while at it, fix the appropriate header files to silence implicit
  declaration of function calls of the GCC compiler.
  |plugins/aiurdemux/aiurcontent.c:582:7: warning: implicit declaration of \
  |                           function 'umask' [-Wimplicit-function-declaration]
  |        umask (0);
  |        ^
  |plugins/aiurdemux/aiurcontent.c:583:11: warning: implicit declaration of \
  |                           function 'mkdir' [-Wimplicit-function-declaration]
  |        if (mkdir (prefix, 0777))

Signed-off-by: Emmanuil Chatzipetru <emmanuil.chatzipetru@nxp.com>