gplay: fix gplay cannot exit when eos
scanf is a block call which will prevent input thread
exit after user input. in previous release, we can kill
input thread using g_thread_unref(), but on l6.6.y release
it cannot work. So change to using poll wait to add
a timeout for stdin reading. So that we can exit the
input thread normally

Signed-off-by: Haihua Hu <jared.hu@nxp.com>